+++ src/main/java/org/dice_group/models/QuatEmbeddingModel.java
@@ -1,7 +1,8 @@
 package org.dice_group.models;
 
-import org.apache.commons.math3.complex.Quaternion;
 import org.dice_group.util.ArrayUtils;
+
+import com.jme3.math.Quaternion;
 
 /**
  * Quaternion-based embedding model
@@ -9,13 +10,13 @@ import org.dice_group.util.ArrayUtils;
  */
 public abstract class QuatEmbeddingModel extends EmbeddingModel {
 
-	protected double[][] relW;
-	protected double[][] relX;
-	protected double[][] relY;
-	protected double[][] relZ;
+	protected float[][] relW;
+	protected float[][] relX;
+	protected float[][] relY;
+	protected float[][] relZ;
 	protected Quaternion[] targetEdge;
 
-	public QuatEmbeddingModel(double[][] relW, double[][] relX, double[][] relY, double[][] relZ) {
+	public QuatEmbeddingModel(float[][] relW, float[][] relX, float[][] relY, float[][] relZ) {
 		this.relW = relW;
 		this.relX = relX;
 		this.relY = relY;
@@ -35,35 +36,35 @@ public abstract class QuatEmbeddingModel extends EmbeddingModel {
 		this.targetEdge = targetEdge;
 	}
 
-	public double[][] getRelW() {
+	public float[][] getRelW() {
 		return relW;
 	}
 
-	public void setRelW(double[][] relW) {
+	public void setRelW(float[][] relW) {
 		this.relW = relW;
 	}
 
-	public double[][] getRelX() {
+	public float[][] getRelX() {
 		return relX;
 	}
 
-	public void setRelX(double[][] relX) {
+	public void setRelX(float[][] relX) {
 		this.relX = relX;
 	}
 
-	public double[][] getRelY() {
+	public float[][] getRelY() {
 		return relY;
 	}
 
-	public void setRelY(double[][] relY) {
+	public void setRelY(float[][] relY) {
 		this.relY = relY;
 	}
 
-	public double[][] getRelZ() {
+	public float[][] getRelZ() {
 		return relZ;
 	}
 
-	public void setRelZ(double[][] relZ) {
+	public void setRelZ(float[][] relZ) {
 		this.relZ = relZ;
 	}
